Fix method resolution order bug affecting calls to overloads in method definitions.
When adding parents to an overload tree, we were not checking to see if the
existing parents were ancestors of the new parent.  Since overloads can gather
parents from multiple levels of the namespace hierarchy, this can happen in
composite namespaces (i.e. class bodies) and can disrupt the method
resolution order because a base class override can be checked before derived
class functions.

Fixing this exposed a bug in alias construction: when flattening an overload
into an alias, we don't want to omit functions whose signature is already
present because we usually don't consider virtual overrides when doing method
resolution and we need the base function to be present.